+++ src/handlers/floats/printdouble/PrintDouble.h
@@ -3,6 +3,6 @@
 
 #include "DoubleUnpacked.h"
 
-char* printDouble(double d, int totalWidth, int fractionalWidth);
+char* printDouble(long double d, int totalWidth, int fractionalWidth);
 
 #endif
